fix: fix set width and height problem and add resize hooks
