+++ packages/s2-core/src/components/sheets/base-sheet/index.tsx
@@ -1,6 +1,6 @@
 import { Event as GEvent } from '@antv/g-canvas';
 import { Pagination, Spin } from 'antd';
-import { debounce, forIn, get, isEmpty, isFunction, merge } from 'lodash';
+import { forIn, get, isEmpty, isFunction, merge } from 'lodash';
 import React, { memo, StrictMode, useEffect, useRef, useState } from 'react';
 import { S2Event } from '@/common/constant';
 import { S2_PREFIX_CLS } from '@/common/constant/classnames';
@@ -27,6 +27,7 @@ import {
   HandleDrillDownIcon,
 } from '@/utils/drill-down/helper';
 import { getBaseCellData } from '@/utils/interaction/formatter';
+import { useResizeEffect } from '@/components/sheets/hooks';
 import './index.less';
 
 export const BaseSheet: React.FC<BaseSheetProps> = memo((props) => {
@@ -61,7 +62,6 @@ export const BaseSheet: React.FC<BaseSheetProps> = memo((props) => {
 
   const [ownSpreadsheet, setOwnSpreadsheet] = useState<SpreadSheet>();
   const [drillFields, setDrillFields] = useState<string[]>([]);
-  const [resizeTimeStamp, setResizeTimeStamp] = useState<number | null>(null);
   const [loading, setLoading] = useState<boolean>(true);
   const [total, setTotal] = useState<number>();
   const [current, setCurrent] = useState<number>(
@@ -234,10 +234,6 @@ export const BaseSheet: React.FC<BaseSheetProps> = memo((props) => {
     update();
   };
 
-  const debounceResize = debounce((e: Event) => {
-    setResizeTimeStamp(e.timeStamp);
-  }, 200);
-
   const renderPagination = (): JSX.Element => {
     const paginationCfg = get(options, 'pagination', false);
     // not show the pagination
@@ -294,28 +290,14 @@ export const BaseSheet: React.FC<BaseSheetProps> = memo((props) => {
 
   useEffect(() => {
     buildSpreadSheet();
-    // 监听窗口变化
-    if (adaptive) window.addEventListener('resize', debounceResize);
     return () => {
       unBindEvent();
       baseSpreadsheet.current.destroy();
-      if (adaptive) window.removeEventListener('resize', debounceResize);
     };
   }, []);
 
-  useEffect(() => {
-    if (!container.current || !ownSpreadsheet) return;
-
-    const style = getComputedStyle(container.current);
-
-    const box = {
-      width: parseInt(style.getPropertyValue('width').replace('px', ''), 10),
-      height: parseInt(style.getPropertyValue('height').replace('px', ''), 10),
-    };
-
-    ownSpreadsheet.changeSize(box?.width, box?.height);
-    ownSpreadsheet.render(false);
-  }, [resizeTimeStamp]);
+  // handle box size change and resize
+  useResizeEffect(container.current, ownSpreadsheet, adaptive, options);
 
   useEffect(() => {
     update(setDataCfg);
